+++ src/codesyncd/websocket/socket_client.ts
@@ -11,7 +11,6 @@ import {
     API_ROUTES,
     STATUS_BAR_MSGS
 } from "../../constants";
-import { getPlanLimitReached } from "../../utils/pricing_utils";
 import { CodeSyncState, CODESYNC_STATES } from "../../utils/state_utils";
 import { setDiffsBeingProcessed } from "../utils";
 
@@ -48,10 +47,6 @@ export class SocketClient {
     }
 
     connect = (canSendDiffs: boolean) => {
-        // Check plan limits
-        const { planLimitReached, canRetry } = getPlanLimitReached();
-		if (planLimitReached && !canRetry) return CodeSyncState.set(CODESYNC_STATES.BUFFER_HANDLER_RUNNING, false);
-
         if (!this.client) {
             this.client = new client();
             (global as any).client = this.client;