+++ src/utils/pricing_utils.ts
@@ -6,89 +6,68 @@ import { CodeSyncState, CODESYNC_STATES } from "./state_utils";
 import { getRepoPlanInfo } from './sync_repo_utils';
 import { generateWebUrl } from './url_utils';
 import { ConfigUtils } from './config_utils';
+import { RepoPlanLimitsUtils } from './repo_utils';
+import { IRepoPlanInfo } from '../interface';
 
 export class PlanLimitsHandler {
 
-	accessToken: string;
+	accessToken: string; // Required
+	repoId: number; // Required
+	repoPath: string; // Optional
+	currentRepoPath: string; // Custom
 	
-	constructor(accessToken: string) {
+	constructor(accessToken: string, repoId: number, repoPath="") {
 		this.accessToken = accessToken;
+		this.repoId = repoId;
+		this.repoPath = repoPath;
+		this.currentRepoPath = pathUtils.getRootPath();
 	}
 
-	async set(repoId: number) {
-		console.log("set", repoId);
-		let repoPath = null;
+	_getRepoPlanInfo = async () : Promise<IRepoPlanInfo> => {
+		// Get Repo Plan Info from server
+		const planInfo = <IRepoPlanInfo>{};
+		planInfo.pricingUrl = generateWebUrl(PRICING_URL_PATH);
+		planInfo.isOrgRepo = false;
+		planInfo.canAvailTrial = false;
+		const json = <any> await getRepoPlanInfo(this.accessToken, this.repoId);
+		if (!json.error) {
+			planInfo.pricingUrl = generateWebUrl("", json.response.url);
+			planInfo.isOrgRepo = json.response.is_org_repo;
+			planInfo.canAvailTrial = json.response.can_avail_trial;
+		}
+		return planInfo;
+	}
+
+	async run() {
+		console.log("planLimitsHandler:run");
 		const configUtils = new ConfigUtils();
 		const config = configUtils.config;
-		if (repoId) {
-			repoPath = configUtils.getRepoPathByRepoId(repoId);
-		} else {
-			// Assuming it is current repo for now
-			repoPath = pathUtils.getRootPath();
-			if (repoPath) {
-				const result = checkSubDir(repoPath);
-				if (result.isSubDir) {
-					repoPath = result.parentRepo;
-				}
-			}
-			const configRepo = config.repos[repoPath];
-			repoId = configRepo.id;
+		if (!config) return;
+		const repoPath = this.repoPath || configUtils.getRepoPathByRepoId(this.repoId);
+		if (!repoPath) return;
+		const repoPlanInfo = await this._getRepoPlanInfo();
+		// Set RepoPlanLimitsState
+		const repoPlanLimitUtils = new RepoPlanLimitsUtils(repoPath);
+		repoPlanLimitUtils.setState(repoPlanInfo.canAvailTrial);
+		// Mark upgradePricingPlan to show button in left panel only if it is currently opened repo
+		if (this.currentRepoPath === repoPath) {
+			vscode.commands.executeCommand('setContext', contextVariables.upgradePricingPlan, true);
+			vscode.commands.executeCommand('setContext', contextVariables.canAvailTrial, repoPlanInfo.canAvailTrial);
+			CodeSyncState.set(CODESYNC_STATES.PRICING_URL, repoPlanInfo.pricingUrl);
+			CodeSyncState.set(CODESYNC_STATES.CAN_AVAIL_TRIAL, repoPlanInfo.canAvailTrial);	
 		}
-		let pricingUrl = generateWebUrl(PRICING_URL_PATH);
-		let isOrgRepo = false;
-		let canAvailTrial = false;
-	
-		// Set time when request is sent
-		CodeSyncState.set(CODESYNC_STATES.REQUEST_SENT_AT, new Date().getTime());
-	
-		const json = <any> await getRepoPlanInfo(this.accessToken, repoId);
-		if (!json.error) {
-			pricingUrl = generateWebUrl("", json.response.url);
-			isOrgRepo = json.response.is_org_repo;
-			canAvailTrial = json.response.can_avail_trial;
-		}
-	
-		// Mark upgradePricingPlan to show button in left panel
-		vscode.commands.executeCommand('setContext', contextVariables.upgradePricingPlan, true);
-		vscode.commands.executeCommand('setContext', contextVariables.canAvailTrial, canAvailTrial);
-	
+		// Create msg for the notification
 		let msg = NOTIFICATION.FREE_TIER_LIMIT_REACHED;
-		const subMsg = isOrgRepo ? NOTIFICATION.UPGRADE_ORG_PLAN : NOTIFICATION.UPGRADE_PRICING_PLAN;
+		const subMsg = repoPlanInfo.isOrgRepo ? NOTIFICATION.UPGRADE_ORG_PLAN : NOTIFICATION.UPGRADE_PRICING_PLAN;
 		msg = `${msg} ${repoPath}. ${subMsg}`;
-		let button = isOrgRepo ? NOTIFICATION_BUTTON.UPGRADE_TO_TEAM : NOTIFICATION_BUTTON.UPGRADE_TO_PRO;
-		if (canAvailTrial) {
-			button = isOrgRepo ? NOTIFICATION_BUTTON.TRY_TEAM_FOR_FREE : NOTIFICATION_BUTTON.TRY_PRO_FOR_FREE;
+		let button = repoPlanInfo.isOrgRepo ? NOTIFICATION_BUTTON.UPGRADE_TO_TEAM : NOTIFICATION_BUTTON.UPGRADE_TO_PRO;
+		if (repoPlanInfo.canAvailTrial) {
+			button = repoPlanInfo.isOrgRepo ? NOTIFICATION_BUTTON.TRY_TEAM_FOR_FREE : NOTIFICATION_BUTTON.TRY_PRO_FOR_FREE;
 		}
-		CodeSyncState.set(CODESYNC_STATES.PRICING_URL, pricingUrl);
-		CodeSyncState.set(CODESYNC_STATES.CAN_AVAIL_TRIAL, canAvailTrial);
 		// Show alert msg
 		vscode.window.showErrorMessage(msg, button).then(async selection => {
 			if (!selection) return;
-			vscode.env.openExternal(vscode.Uri.parse(pricingUrl));
+			vscode.env.openExternal(vscode.Uri.parse(repoPlanInfo.pricingUrl));
 		});
 	}
 }
-
-export const getPlanLimitReached = () => {
-	/*
-		If pricingAlertlock is acquried by any IDE instance, means pricing limit has been reached. 
-	*/
-	const planLimitReached = false;
-	const requestSentAt = CodeSyncState.get(CODESYNC_STATES.REQUEST_SENT_AT);
-	const canRetry = requestSentAt && (new Date().getTime() - requestSentAt) > RETRY_REQUEST_AFTER;
-	return {
-		planLimitReached,
-		canRetry
-	};
-};
-
-export const resetPlanLimitReached = () => {
-	/*
-		Checks from server if repo is a User's repo or an Organization's repo
-		- Sets alert msg and pricing URL accordingly
-		- Sets REQUEST_SENT_AT in CodeSyncState after which we retry syncing data
-	*/
-	vscode.commands.executeCommand('setContext', contextVariables.upgradePricingPlan, false);
-	// Set time to "" when request is sent
-	CodeSyncState.set(CODESYNC_STATES.REQUEST_SENT_AT, "");
-};