+++ src/utils/state_utils.ts
@@ -10,7 +10,6 @@ export const CODESYNC_STATES = {
     POPULATE_BUFFER_RUNNING: "populateBufferRunning",
     BUFFER_HANDLER_RUNNING: "bufferHandlerRunning",
     PRICING_URL: "pricingUrl",
-    REQUEST_SENT_AT: "requestSentAt",
     SYNCING_BRANCH: "syncingBranch",
     IS_SYNCING_BRANCH: "isSyncingBranch",
     UPLOADING_TO_S3: "uploadingToS3",