+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dropping a log when we reset the config file
 - Updated messages for a disconnected repository in menus, left panel and notifications
 - Repo based plan limits implemented, improved button text to upgrade pricing plans, removed pricing_lock file
+- Skipping Diffs for the repository for which limit is reached. We try sending diffs after allowed wait time
 
 ## [3.38.0] - 2024-01-12
 - Sending commit_hash for new files